Assign correct packages to classes
