--- jetty/src/test/java/com/github/valb3r/letsencrypthelper/JettyExpiredKeyPebbleTest.java
@@ -1,4 +1,0 @@
-package com.github.valb3r.letsencrypthelper;
-
-class JettyExpiredKeyPebbleTest extends ExpiredKeyPebbleTest {
-}
