--- jetty/src/test/java/com/github/valb3r/letsencrypthelper/JettyNoKeystorePebbleTest.java
@@ -1,4 +1,0 @@
-package com.github.valb3r.letsencrypthelper;
-
-class JettyNoKeystorePebbleTest extends NoKeystorePebbleTest {
-}
