--- jetty/src/test/java/com/github/valb3r/letsencrypthelper/JettyNotExpiredKeystorePebbleTest.java
@@ -1,4 +1,0 @@
-package com.github.valb3r.letsencrypthelper;
-
-class JettyNotExpiredKeystorePebbleTest extends NotExpiredKeystorePebbleTest {
-}
