+++ jetty/src/test/java/com/github/valb3r/letsencrypthelper/jetty/JettyNoKeystorePebbleTest.java
@@ -1,0 +1,6 @@
+package com.github.valb3r.letsencrypthelper.jetty;
+
+import com.github.valb3r.letsencrypthelper.NoKeystorePebbleTest;
+
+class JettyNoKeystorePebbleTest extends NoKeystorePebbleTest {
+}
