+++ jetty/src/test/java/com/github/valb3r/letsencrypthelper/jetty/JettyNotExpiredKeystorePebbleTest.java
@@ -1,0 +1,6 @@
+package com.github.valb3r.letsencrypthelper.jetty;
+
+import com.github.valb3r.letsencrypthelper.NotExpiredKeystorePebbleTest;
+
+class JettyNotExpiredKeystorePebbleTest extends NotExpiredKeystorePebbleTest {
+}
