+++ jetty/src/test/java/com/github/valb3r/letsencrypthelper/jetty/dummyapp/DummyApp.java
@@ -1,12 +1,12 @@
-package com.github.valb3r.letsencrypthelper.dummyapp;
+package com.github.valb3r.letsencrypthelper.jetty.dummyapp;
 
-import com.github.valb3r.letsencrypthelper.TomcatWellKnownLetsEncryptChallengeEndpointConfig;
+import com.github.valb3r.letsencrypthelper.jetty.JettyWellKnownLetsEncryptChallengeEndpointConfig;
 import org.springframework.boot.SpringApplication;
 import org.springframework.boot.autoconfigure.SpringBootApplication;
 import org.springframework.context.annotation.Import;
 
 @SpringBootApplication
-@Import({TomcatWellKnownLetsEncryptChallengeEndpointConfig.class})
+@Import({JettyWellKnownLetsEncryptChallengeEndpointConfig.class})
 public class DummyApp {
 
     public static void main(String[] args) {
