--- tomcat/src/test/java/com/github/valb3r/letsencrypthelper/TomcatNoKeystorePebbleTest.java
@@ -1,4 +1,0 @@
-package com.github.valb3r.letsencrypthelper;
-
-class TomcatNoKeystorePebbleTest extends NoKeystorePebbleTest {
-}
