+++ tomcat/src/test/java/com/github/valb3r/letsencrypthelper/tomcat/TomcatNotExpiredKeystorePebbleTest.java
@@ -1,0 +1,6 @@
+package com.github.valb3r.letsencrypthelper.tomcat;
+
+import com.github.valb3r.letsencrypthelper.NotExpiredKeystorePebbleTest;
+
+class TomcatNotExpiredKeystorePebbleTest extends NotExpiredKeystorePebbleTest {
+}
