+++ tomcat/src/test/java/com/github/valb3r/letsencrypthelper/tomcat/dummyapp/DummyApp.java
@@ -1,12 +1,12 @@
-package com.github.valb3r.letsencrypthelper.dummyapp;
+package com.github.valb3r.letsencrypthelper.tomcat.dummyapp;
 
-import com.github.valb3r.letsencrypthelper.JettyWellKnownLetsEncryptChallengeEndpointConfig;
+import com.github.valb3r.letsencrypthelper.tomcat.TomcatWellKnownLetsEncryptChallengeEndpointConfig;
 import org.springframework.boot.SpringApplication;
 import org.springframework.boot.autoconfigure.SpringBootApplication;
 import org.springframework.context.annotation.Import;
 
 @SpringBootApplication
-@Import({JettyWellKnownLetsEncryptChallengeEndpointConfig.class})
+@Import({TomcatWellKnownLetsEncryptChallengeEndpointConfig.class})
 public class DummyApp {
 
     public static void main(String[] args) {
